requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
psycopg2-binary
python-dotenv==1.0.1
PyJWT==2.8.0
pandas
openpyxl
numpy
scipy
gunicorn
eventlet
qrcode[pil]==7.4.2
argon2-cffi
passlib
flask-limiter
redis
Flask-Talisman
google-cloud-secret-manager